+++ activerecord/lib/active_record/connection_adapters/sqlite3_adapter.rb
@@ -76,22 +76,15 @@ module ActiveRecord
         json:         { name: "json" },
       }
 
-      ##
-      # :singleton-method:
-      # Indicates whether boolean values are stored in sqlite3 databases as 1
-      # and 0 or 't' and 'f'. Leaving <tt>ActiveRecord::ConnectionAdapters::SQLite3Adapter.represent_boolean_as_integer</tt>
-      # set to false is deprecated. SQLite databases have used 't' and 'f' to
-      # serialize boolean values and must have old data converted to 1 and 0
-      # (its native boolean serialization) before setting this flag to true.
-      # Conversion can be accomplished by setting up a rake task which runs
-      #
-      #   ExampleModel.where("boolean_column = 't'").update_all(boolean_column: 1)
-      #   ExampleModel.where("boolean_column = 'f'").update_all(boolean_column: 0)
-      # for all models and all boolean columns, after which the flag must be set
-      # to true by adding the following to your <tt>application.rb</tt> file:
-      #
-      #   Rails.application.config.active_record.sqlite3.represent_boolean_as_integer = true
-      class_attribute :represent_boolean_as_integer, default: false
+      def self.represent_boolean_as_integer=(value) # :nodoc:
+        if value == false
+          raise "`.represent_boolean_as_integer=` is now always true, so make sure your application can work with it and remove this settings."
+        end
+
+        ActiveSupport::Deprecation.warn(
+          "`.represent_boolean_as_integer=` is now always true, so setting this is deprecated and will be removed in Rails 6.1."
+        )
+      end
 
       class StatementPool < ConnectionAdapters::StatementPool # :nodoc:
         private
@@ -314,11 +307,6 @@ module ActiveRecord
         rename_table_indexes(table_name, new_name)
       end
 
-      def valid_alter_table_type?(type, options = {})
-        !invalid_alter_table_type?(type, options)
-      end
-      deprecate :valid_alter_table_type?
-
       def add_column(table_name, column_name, type, options = {}) #:nodoc:
         if invalid_alter_table_type?(type, options)
           alter_table(table_name) do |definition|
@@ -388,14 +376,6 @@ module ActiveRecord
           }
           ForeignKeyDefinition.new(table_name, row["table"], options)
         end
-      end
-
-      def insert_fixtures(rows, table_name)
-        ActiveSupport::Deprecation.warn(<<-MSG.squish)
-          `insert_fixtures` is deprecated and will be removed in the next version of Rails.
-          Consider using `insert_fixtures_set` for performance improvement.
-        MSG
-        insert_fixtures_set(table_name => rows)
       end
 
       def insert_fixtures_set(fixture_set, tables_to_delete = [])