+++ railties/lib/rails/command/environment_argument.rb
@@ -8,23 +8,13 @@ module Rails
       extend ActiveSupport::Concern
 
       included do
-        argument :environment, optional: true, banner: "environment"
-
         class_option :environment, aliases: "-e", type: :string,
           desc: "Specifies the environment to run this console under (test/development/production)."
       end
 
       private
         def extract_environment_option_from_argument
-          if environment
-            self.options = options.merge(environment: acceptable_environment(environment))
-
-            ActiveSupport::Deprecation.warn "Passing the environment's name as a " \
-                                            "regular argument is deprecated and "  \
-                                            "will be removed in the next Rails "   \
-                                            "version. Please, use the -e option "  \
-                                            "instead."
-          elsif options[:environment]
+          if options[:environment]
             self.options = options.merge(environment: acceptable_environment(options[:environment]))
           else
             self.options = options.merge(environment: Rails::Command.environment)