+++ railties/test/application/console_test.rb
@@ -149,7 +149,7 @@ class FullStackConsoleTest < ActiveSupport::TestCase
   end
 
   def test_environment_option_and_irb_option
-    spawn_console("test -- --verbose")
+    spawn_console("-e test -- --verbose")
 
     write_prompt "a = 1", "a = 1"
     write_prompt "puts Rails.env", "puts Rails.env\r\ntest"